+++ src/calmjs/testing/spec.py
@@ -29,3 +29,24 @@ def advice_order(spec, extras):
             raise toolchain.ToolchainAbort()  # pragma: no cover
 
     spec.advise(toolchain.CLEANUP, verify_build_dir)
+
+
+def advice_marker(spec, extras):
+    # this documents the behavior if the implementation of a toolchain
+    # advice that was registered had acted immediately on the spec
+    from calmjs import toolchain
+
+    spec.setdefault('marker_too_soon', [])
+    spec['marker_too_soon'].append(
+        (list(spec.get('advice_packages', [])), extras)
+    )
+
+    # to properly manage the process, it should register a function as
+    # an advice.
+    def advice():
+        spec.setdefault('marker_delayed', [])
+        spec['marker_delayed'].append(
+            (list(spec.get('advice_packages', [])), extras)
+        )
+
+    spec.advise(toolchain.SETUP, advice)