+++ src/calmjs/tests/test_runtime.py
@@ -799,6 +799,104 @@ class ToolchainRuntimeTestCase(unittest.TestCase):
             result = rt(['--export-target', 'dummy'])
         self.assertNotIn('dummy', result)
 
+    def test_spec_toolchain_advice_apply(self):
+        # This mostly just run through the toolchain to ensure that the
+        # advice is applied without any arguments.
+        from calmjs.registry import _inst as root_registry
+        key_t_a = toolchain.CALMJS_TOOLCHAIN_ADVICE
+        key_t_a_a = toolchain.CALMJS_TOOLCHAIN_ADVICE_APPLY
+        stub_stdouts(self)
+
+        self.addCleanup(root_registry.records.pop, key_t_a, None)
+        self.addCleanup(root_registry.records.pop, key_t_a_a, None)
+
+        make_dummy_dist(self, ((
+            'entry_points.txt',
+            '[calmjs.toolchain.advice]\n'
+            'calmjs.toolchain:NullToolchain'
+            ' = calmjs.testing.spec:advice_marker\n'
+            '\n'
+            '[calmjs.toolchain.advice.apply]\n'
+            'example.package = example.package\n'
+        ),), 'example.package', '1.0')
+
+        make_dummy_dist(self, ((
+            'entry_points.txt',
+            '[calmjs.toolchain.advice.apply]\n'
+            'example.package = example.package[argument]\n'
+        ),), 'example.other_package', '1.0')
+
+        make_dummy_dist(self, (
+            ('requires.txt', '\n'.join([
+                'example.package',
+                'example.other_package',
+            ])),
+            ('entry_points.txt', '',),
+        ), 'dependent', '1.0')
+
+        working_set = pkg_resources.WorkingSet([self._calmjs_testing_tmpdir])
+
+        root_registry.records[key_t_a] = toolchain.AdviceRegistry(
+            key_t_a, _working_set=working_set)
+        root_registry.records[key_t_a_a] = toolchain.AdviceApplyRegistry(
+            key_t_a_a, _working_set=working_set)
+
+        tc = toolchain.NullToolchain()
+        rt = runtime.SourcePackageToolchainRuntime(tc)
+
+        result = rt([
+            '--export-target', join(self.cwd, 'dummy_export'),
+            'example.package', '-v',
+        ])
+
+        self.assertEqual([([], [])], result['marker_too_soon'])
+        self.assertEqual([(['example.package'], [])], result['marker_delayed'])
+        err = sys.stderr.getvalue()
+        err_lines = err.splitlines()
+
+        self.assertIn(
+            "source package 'example.package' specified advice packages",
+            err_lines[0])
+        self.assertIn('example.package', err_lines[0])
+        self.assertIn('to be applied', err_lines[0])
+
+        # Doing it normally should not result in that optional key.
+        with pretty_logging(logger='calmjs', stream=mocks.StringIO()) as s:
+            result = rt([
+                '--export-target', join(self.cwd, 'dummy_export'),
+                'example.other_package', '-v',
+            ])
+
+        # arguments will be passed, but the advice_packages key won't be
+        # assigned.
+        self.assertEqual([([], ['argument'])], result['marker_too_soon'])
+        # naturally, it would be available in the setup.
+        self.assertEqual(
+            [(['example.package[argument]'], ['argument'])],
+            result['marker_delayed']
+        )
+
+        err_lines = s.getvalue().splitlines()
+        self.assertIn(
+            "source package 'example.other_package' specified advice packages",
+            err_lines[0])
+        self.assertIn('example.package[argument]', err_lines[0])
+        self.assertIn('to be applied', err_lines[0])
+
+        # This tests the dependent case where the dependent package
+        # declares both as dependents, but it doesn't declare the
+        # records for the relevant registries.
+        with pretty_logging(logger='calmjs', stream=mocks.StringIO()) as s:
+            result = rt([
+                '--export-target', join(self.cwd, 'dummy_export'),
+                'dependent', '-v',
+            ])
+
+        # shouldn't have anything.
+        self.assertNotIn('marker_too_soon', result)
+        self.assertNotIn('marker_delayed', result)
+        self.assertEqual('', s.getvalue())
+
     def test_spec_advise_debugger(self):
         # this is meant for advanced usage, thus undocumented.
         from calmjs import utils