+++ CHANGES.rst
@@ -15,6 +15,12 @@ Changelog
   enabled for all subclasses of Runtime.  [
   `#55 <https://github.com/calmjs/calmjs/issues/55>`_
   ]
+- Provide a registry for packages to specify which requirement (package)
+  to acquire toolchain advice from to apply by default for the relevant
+  toolchains.  This also necessitated some changes to where the optional
+  advices are applied, from the previous location applied as a setup
+  level advice applied by the standard toolchain runtime, to the default
+  sequence within the default toolchain itself.
 
 3.3.1 (2018-08-20)
 ------------------